refactor: change index.js to index.ts
