+++ .env.example
@@ -1,0 +1,2 @@
+HTTP_PORT=8181
+WEBSOCKET_PORT=3000
